# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fleet Management API",
    "version": "1.0.0"
  },
  "paths": {
    "/vehicles": {
      "get": {
        "summary": "Get list of all vehicles",
        "responses": {
          "200": {
            "description": "List of vehicles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IVehicle"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new vehicle",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vehicle created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IVehicle"
                }
              }
            }
          }
        }
      }
    },
    "/vehicles/{vehicleId}/maintenance": {
      "post": {
        "summary": "Add maintenance record to a vehicle",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMaintenanceRecordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Maintenance record added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IMaintenanceRecord"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get maintenance records of a vehicle",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IMaintenanceRecord"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/vehicles/{vehicleId}/status": {
      "get": {
        "summary": "Get real-time status of a vehicle",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IVehicleStatus"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateVehicleDto": {
        "type": "object",
        "properties": {
          "make": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          },
          "vehicleModel": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          },
          "year": {
            "type": "integer",
            "minimum": 1900,
            "maximum": 2024
          },
          "ecuDeviceId": {
            "type": "string",
            "minLength": 5,
            "maxLength": 50
          }
        },
        "required": ["make", "vehicleModel", "year", "ecuDeviceId"]
      },
      "CreateMaintenanceRecordDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "cost": {
            "type": "number"
          }
        },
        "required": ["description", "date", "cost"]
      },
      "IVehicle": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "make": {
            "type": "string"
          },
          "vehicleModel": {
            "type": "string"
          },
          "year": {
            "type": "integer"
          },
          "ecuDeviceId": {
            "type": "string"
          },
          "aggregatedSensorData": {
            "$ref": "#/components/schemas/AggregatedSensorData"
          },
          "lastMaintenanceRecord": {
            "$ref": "#/components/schemas/CreateMaintenanceRecordDto"
          },
          "usageAnalytics": {
            "$ref": "#/components/schemas/UsageAnalytics"
          }
        }
      },
      "IMaintenanceRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "cost": {
            "type": "number"
          },
          "vehicleId": {
            "type": "string"
          }
        }
      },
      "IVehicleStatus": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "aggregatedSensorData": {
            "$ref": "#/components/schemas/AggregatedSensorData"
          },
          "lastMaintenanceRecord": {
            "$ref": "#/components/schemas/CreateMaintenanceRecordDto"
          },
          "usageAnalytics": {
            "$ref": "#/components/schemas/UsageAnalytics"
          }
        }
      },
      "AggregatedSensorData": {
        "type": "object",
        "properties": {
          "odometerReading": {
            "type": "number"
          },
          "engineHours": {
            "type": "number"
          },
          "fuelLevel": {
            "type": "string"
          },
          "lastGpsLocation": {
            "type": "object",
            "properties": {
              "latitude": {
                "type": "number"
              },
              "longitude": {
                "type": "number"
              },
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UsageAnalytics": {
        "type": "object",
        "properties": {
          "distanceTraveled": {
            "type": "number"
          },
          "hoursOperated": {
            "type": "number"
          }
        }
      }
    }
  }
}